requirements.txt
# RAG Traditional - Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
openai>=1.10.0

# Vector Store
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.3.1
tiktoken>=0.5.2

# Data Processing
pandas>=2.1.4
numpy>=1.26.3

# Environment Management
python-dotenv>=1.0.0

# Optional: Alternative LLM providers
anthropic>=0.18.0

# Free/Local LLM
ollama>=0.1.0

# Neo4j for Knowledge Graph
neo4j>=5.15.0
py2neo>=2021.2.4

# NLP for label detection
spacy>=3.7.0

# Jupyter for demos
jupyter>=1.0.0
ipykernel>=6.29.0
notebook>=7.0.0

# Utilities
tqdm>=4.66.1
